Refactoring fo envMultiply and envConst chunks
