+++ src/scene/shader-lib/chunks-wgsl/chunks-wgsl.js
@@ -35,8 +35,7 @@ import decodePS from './common/frag/decode.js';
 // import encodePS from './common/frag/encode.js';
 // import endPS from './lit/frag/end.js';
 import envAtlasPS from './common/frag/envAtlas.js';
-// import envConstPS from './common/frag/envConst.js';
-import envMultiplyPS from './common/frag/envMultiply.js';
+import envProcPS from './common/frag/envProc.js';
 // import falloffInvSquaredPS from './lit/frag/falloffInvSquared.js';
 // import falloffLinearPS from './lit/frag/falloffLinear.js';
 // import floatUnpackingPS from './lit/frag/float-unpacking.js';
@@ -240,7 +239,7 @@ const shaderChunksWGSL = {
     // endPS,
     envAtlasPS,
     // envConstPS,
-    envMultiplyPS,
+    envProcPS,
     // falloffInvSquaredPS,
     // falloffLinearPS,
     // floatUnpackingPS,
