--- src/scene/shader-lib/chunks-wgsl/common/frag/envMultiply.js
@@ -1,7 +1,0 @@
-export default /* wgsl */`
-uniform skyboxIntensity : f32;
-
-fn processEnvironment(color : vec3f) -> vec3f {
-    return color * uniform.skyboxIntensity;
-}
-`;
