+++ src/scene/shader-lib/chunks-wgsl/common/frag/envProc.js
@@ -1,0 +1,13 @@
+export default /* wgsl */`
+#ifdef LIT_SKYBOX_INTENSITY
+    uniform skyboxIntensity : f32;
+#endif
+
+fn processEnvironment(color : vec3f) -> vec3f {
+    #ifdef LIT_SKYBOX_INTENSITY
+        return color * uniform.skyboxIntensity;
+    #else
+        return color;
+    #endif
+}
+`;
