--- src/scene/shader-lib/chunks/common/frag/envConst.js
@@ -1,5 +1,0 @@
-export default /* glsl */`
-vec3 processEnvironment(vec3 color) {
-    return color;
-}
-`;
