--- src/scene/shader-lib/chunks/common/frag/envMultiply.js
@@ -1,7 +1,0 @@
-export default /* glsl */`
-uniform float skyboxIntensity;
-
-vec3 processEnvironment(vec3 color) {
-    return color * skyboxIntensity;
-}
-`;
