+++ src/scene/shader-lib/chunks/common/frag/envProc.js
@@ -1,0 +1,13 @@
+export default /* glsl */`
+#ifdef LIT_SKYBOX_INTENSITY
+    uniform float skyboxIntensity;
+#endif
+
+vec3 processEnvironment(vec3 color) {
+    #ifdef LIT_SKYBOX_INTENSITY
+        return color * skyboxIntensity;
+    #else
+        return color;
+    #endif
+}
+`;
