+++ src/scene/shader-lib/chunks/skybox/frag/skybox.js
@@ -1,5 +1,7 @@
 export default /* glsl */`
-    #include "envMultiplyPS"
+    #define LIT_SKYBOX_INTENSITY
+
+    #include "envProcPS"
     #include "gammaPS"
     #include "tonemappingPS"
